feat: 길드 유니크 id 서버에 추가 로직
