+++ src/server/view/server/page.tsx
@@ -1,24 +1,12 @@
-import { useEffect } from "react";
+import { useGuildIdListener } from "./hook/use_guild_id_listener";
 import ServerSection from "./section/server_section";
 import TitleSection from "./section/title_section";
-import { toast, ToastContainer } from "react-toastify";
+import { ToastContainer } from "react-toastify";
 import "react-toastify/dist/ReactToastify.css";
 
 export default function ServerPage() {
-  useEffect(() => {
-    const handleMessage = (event: MessageEvent) => {
-      const { status, message } = event.data;
+  useGuildIdListener(); // 리브 봇 추가 창에서 guildId가 저장이 되면 해당 guildId를 관리자서버에 전송
 
-      if (status === "success") {
-        toast.success(message);
-      } else if (status === "error") {
-        toast.error(message);
-      }
-    };
-
-    window.addEventListener("message", handleMessage);
-    return () => window.removeEventListener("message", handleMessage);
-  }, []);
   return (
     <>
       <TitleSection />
